# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Steamworks Publisher Stats Exporter API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface ConnectionCredentials {
  /** Steam sessionid cookie value */
  sessionid: string;
  /** Steam steamLoginSecure cookie value */
  steamLoginSecure: string;
}

export interface ConnectionResult {
  /** Publisher account name */
  publisherName: string;
  /** Number of base games (excluding demos/playtests) */
  gameCount: number;
}

export interface GameInfo {
  /** Steam App ID */
  appId: number;
  /** Game name */
  name: string;
  /** App type (game, dlc, etc.) */
  type: string;
}

export interface GamesListResult {
  games: GameInfo[];
  /** Apps excluded (demos, playtests) */
  skipped: GameInfo[];
}

/**
 * Time granularity for the pull
 */
export type PullRequestGranularity =
  (typeof PullRequestGranularity)[keyof typeof PullRequestGranularity];

export const PullRequestGranularity = {
  daily: "daily",
  weekly: "weekly",
  monthly: "monthly",
  lifetime: "lifetime",
} as const;

export interface PullRequest {
  sessionid: string;
  steamLoginSecure: string;
  /** App IDs to pull data for */
  appIds: number[];
  /** Time granularity for the pull */
  granularity: PullRequestGranularity;
}

export interface PullStartResult {
  /** Unique job identifier for polling */
  jobId: string;
}

export interface PullProgressItem {
  gameIndex: number;
  totalGames: number;
  gameName: string;
  metric: string;
  estimatedSecondsRemaining?: number;
}

export type PullStatusStatus =
  (typeof PullStatusStatus)[keyof typeof PullStatusStatus];

export const PullStatusStatus = {
  running: "running",
  completed: "completed",
  cancelled: "cancelled",
  failed: "failed",
  session_expired: "session_expired",
} as const;

export type PullStatusGameErrorsItem = {
  appId: number;
  gameName: string;
  error: string;
};

export interface PullStatus {
  status: PullStatusStatus;
  progress?: PullProgressItem;
  errorMessage?: string;
  gameErrors?: PullStatusGameErrorsItem[];
}