# lib/api-zod/src/generated/types/connectionCredentials.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Steamworks Publisher Stats Exporter API
 * OpenAPI spec version: 0.1.0
 */

export interface ConnectionCredentials {
  /** Steam sessionid cookie value */
  sessionid: string;
  /** Steam steamLoginSecure cookie value */
  steamLoginSecure: string;
}