# lib/api-zod/src/generated/types/connectionResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Steamworks Publisher Stats Exporter API
 * OpenAPI spec version: 0.1.0
 */

export interface ConnectionResult {
  /** Publisher account name */
  publisherName: string;
  /** Number of base games (excluding demos/playtests) */
  gameCount: number;
}