# lib/api-zod/src/generated/types/gameInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Steamworks Publisher Stats Exporter API
 * OpenAPI spec version: 0.1.0
 */

export interface GameInfo {
  /** Steam App ID */
  appId: number;
  /** Game name */
  name: string;
  /** App type (game, dlc, etc.) */
  type: string;
}