# lib/api-zod/src/generated/types/gamesListResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Steamworks Publisher Stats Exporter API
 * OpenAPI spec version: 0.1.0
 */
import type { GameInfo } from "./gameInfo";

export interface GamesListResult {
  games: GameInfo[];
  /** Apps excluded (demos, playtests) */
  skipped: GameInfo[];
}