# lib/api-zod/src/generated/types/pullRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Steamworks Publisher Stats Exporter API
 * OpenAPI spec version: 0.1.0
 */
import type { PullRequestGranularity } from "./pullRequestGranularity";

export interface PullRequest {
  sessionid: string;
  steamLoginSecure: string;
  /** App IDs to pull data for */
  appIds: number[];
  /** Time granularity for the pull */
  granularity: PullRequestGranularity;
}