# lib/api-zod/src/generated/types/pullStartResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Steamworks Publisher Stats Exporter API
 * OpenAPI spec version: 0.1.0
 */

export interface PullStartResult {
  /** Unique job identifier for polling */
  jobId: string;
}