# lib/api-zod/src/generated/types/pullStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Steamworks Publisher Stats Exporter API
 * OpenAPI spec version: 0.1.0
 */
import type { PullProgressItem } from "./pullProgressItem";
import type { PullStatusGameErrorsItem } from "./pullStatusGameErrorsItem";
import type { PullStatusStatus } from "./pullStatusStatus";

export interface PullStatus {
  status: PullStatusStatus;
  progress?: PullProgressItem;
  errorMessage?: string;
  gameErrors?: PullStatusGameErrorsItem[];
}